+++ lib/which.php
@@ -2,27 +2,21 @@
 declare(strict_types=1);
 namespace Which;
 
-use Rx\{Observable};
-
 /**
  * Finds the first instance of an executable in the system path.
  * @param string $command The command to be resolved.
  * @param bool $all Value indicating whether to return all executables found, instead of just the first one.
  * @param array $options The options to be passed to the finder.
- * @return Observable A string, or an array of strings, specifying the path(s) of the found executable(s).
+ * @return string|string[] A string, or an array of strings, specifying the path(s) of the found executable(s).
+ * @throws \RuntimeException The command was not found.
  */
-function which(string $command, bool $all = false, array $options = []): Observable {
-  $finder = new Finder(
+function which(string $command, bool $all = false, array $options = []) {
+  $executables = (new Finder(
     $options['path'] ?? '',
     $options['extensions'] ?? '',
     $options['pathSeparator'] ?? ''
-  );
+  ))->find($command);
 
-  $executables = $finder->find($command);
-  return $executables
-    ->isEmpty()
-    ->flatMap(function($isEmpty) use ($all, $command, $executables) {
-      if ($isEmpty) return Observable::error(new \RuntimeException("Command not found: $command"));
-      return $all ? $executables->toArray() : $executables->take(1);
-    });
+  if (!$executables) throw new \RuntimeException("Command not found: $command");
+  return $all ? $executables : $executables[0];
 }